MDE/PKFE-6 component restructure
