+++ app/front-end/src/features/editor/components/fileTreeView/fileTreeItem/fileTreeItem.tsx
@@ -1,0 +1,128 @@
+import { getIconFromFileType, isExpandable } from '@/features/editor/utils';
+import FolderRounded from '@mui/icons-material/FolderRounded';
+import Collapse from '@mui/material/Collapse';
+import { alpha, styled } from '@mui/material/styles';
+import { TransitionProps } from '@mui/material/transitions';
+import { TreeItem2Icon, TreeItem2Provider } from '@mui/x-tree-view';
+import { treeItemClasses } from '@mui/x-tree-view/TreeItem';
+import { TreeItem2Checkbox, TreeItem2Content, TreeItem2IconContainer, TreeItem2Root } from '@mui/x-tree-view/TreeItem2';
+import { TreeItem2DragAndDropOverlay } from '@mui/x-tree-view/TreeItem2DragAndDropOverlay';
+import { unstable_useTreeItem2 as useTreeItem2, UseTreeItem2Parameters } from '@mui/x-tree-view/useTreeItem2';
+import { animated, useSpring } from '@react-spring/web';
+import clsx from 'clsx';
+import React from 'react';
+import { FileTreeItemLabel } from '.';
+
+const StyledFileTreeItemRoot = styled(TreeItem2Root)(({ theme }) => ({
+  color: theme.palette.mode === 'light' ? theme.palette.grey[800] : theme.palette.grey[400],
+  position: 'relative',
+  [`& .${treeItemClasses.groupTransition}`]: {
+    marginLeft: theme.spacing(3.5),
+  },
+})) as unknown as typeof TreeItem2Root;
+
+const StyledFileTreeItemContent = styled(TreeItem2Content)(({ theme }) => ({
+  flexDirection: 'row-reverse',
+  borderRadius: theme.spacing(0.7),
+  marginBottom: theme.spacing(0.5),
+  marginTop: theme.spacing(0.5),
+  padding: theme.spacing(0.5),
+  paddingRight: theme.spacing(1),
+  fontWeight: 500,
+  ['&.Mui-expanded ']: {
+    '&:not(.Mui-focused, .Mui-selected, .Mui-selected.Mui-focused) .labelIcon': {
+      color: theme.palette.mode === 'light' ? theme.palette.primary.main : theme.palette.primary.dark,
+    },
+    '&::before': {
+      content: '""',
+      display: 'block',
+      position: 'absolute',
+      left: '16px',
+      top: '44px',
+      height: 'calc(100% - 48px)',
+      width: '1.5px',
+      backgroundColor: theme.palette.mode === 'light' ? theme.palette.grey[300] : theme.palette.grey[700],
+    },
+  },
+  '&:hover': {
+    backgroundColor: alpha(theme.palette.primary.main, 0.1),
+    color: theme.palette.mode === 'light' ? theme.palette.primary.main : 'white',
+  },
+  ['&.Mui-focused, &.Mui-selected, &.Mui-selected.Mui-focused']: {
+    backgroundColor: theme.palette.mode === 'light' ? theme.palette.primary.main : theme.palette.primary.dark,
+    color: theme.palette.primary.contrastText,
+  },
+}));
+
+const TransitionComponent = (props: TransitionProps) => {
+  const AnimatedCollapse = animated(Collapse);
+  const style = useSpring({
+    to: {
+      opacity: props.in ? 1 : 0,
+      transform: `translate3d(0,${props.in ? 0 : 20}px,0)`,
+    },
+  });
+
+  return <AnimatedCollapse style={style} {...props} />;
+};
+
+interface FileTreeItemProps
+  extends Omit<UseTreeItem2Parameters, 'rootRef'>,
+    Omit<React.HTMLAttributes<HTMLLIElement>, 'onFocus'> {}
+
+export const FileTreeItem = React.forwardRef(function CustomTreeItem(
+  props: FileTreeItemProps,
+  ref: React.Ref<HTMLLIElement>
+) {
+  const { id, itemId, label, disabled, children, ...other } = props;
+
+  const {
+    getRootProps,
+    getContentProps,
+    getIconContainerProps,
+    getCheckboxProps,
+    getLabelProps,
+    getGroupTransitionProps,
+    getDragAndDropOverlayProps,
+    status,
+    publicAPI,
+  } = useTreeItem2({ id, itemId, children, label, disabled, rootRef: ref });
+
+  const item = publicAPI.getItem(itemId);
+  const expandable = isExpandable(children);
+  let icon;
+  if (expandable) {
+    icon = FolderRounded;
+  } else if (item.fileType) {
+    icon = getIconFromFileType(item.fileType);
+  }
+
+  const handleClick = (fileName: string, fileID: string) => {
+    console.log(`Clicked on ${fileName} (${fileID})`); // TODO: implement file fetching from back-end
+  };
+
+  return (
+    <TreeItem2Provider itemId={itemId}>
+      <StyledFileTreeItemRoot {...getRootProps(other)} onClick={() => handleClick(item.label, item.id)}>
+        <StyledFileTreeItemContent
+          {...getContentProps({
+            className: clsx('content', {
+              'Mui-expanded': status.expanded,
+              'Mui-selected': status.selected,
+              'Mui-focused': status.focused,
+              'Mui-disabled': status.disabled,
+            }),
+          })}
+        >
+          <TreeItem2IconContainer {...getIconContainerProps()}>
+            <TreeItem2Icon status={status} />
+          </TreeItem2IconContainer>
+          <TreeItem2Checkbox {...getCheckboxProps()} />
+          <FileTreeItemLabel {...getLabelProps({ icon, expandable: expandable && status.expanded })} />
+          <TreeItem2DragAndDropOverlay {...getDragAndDropOverlayProps()} />
+        </StyledFileTreeItemContent>
+        {children && <TransitionComponent {...getGroupTransitionProps()} />}
+      </StyledFileTreeItemRoot>
+    </TreeItem2Provider>
+  );
+});
