+++ app/front-end/src/features/editor/components/fileTreeView/fileTreeItem/index.ts
@@ -1,0 +1,2 @@
+export { FileTreeItem } from './fileTreeItem';
+export { FileTreeItemLabel } from './fileTreeItemLabel';
