+++ app/front-end/src/features/editor/components/index.ts
@@ -1,4 +1,4 @@
-export { Hierarchy } from './hierarchy';
+export { FileTreeView } from './fileTreeView/fileTreeView';
 export { Toolbar } from './toolbar';
 export { Editor } from './editor';
 export { Filebar } from './filebar';
