+++ app/front-end/src/features/editor/mockData/index.ts
@@ -1,0 +1,1 @@
+export { FileTreeViewMockData } from './fileTreeViewMockData';
