+++ app/front-end/src/features/editor/types/fileTreeViewItemProps.ts
@@ -1,0 +1,7 @@
+import { FileTypes } from '@/types/enums';
+
+export type FileTreeViewItemProps = {
+  fileType?: FileTypes;
+  id: string;
+  label: string;
+};
