+++ app/front-end/src/features/editor/types/index.ts
@@ -1,0 +1,1 @@
+export type { FileTreeViewItemProps } from './fileTreeViewItemProps';
