+++ app/front-end/src/features/editor/utils/index.ts
@@ -1,0 +1,1 @@
+export { getIconFromFileType, isExpandable } from './helpers';
