+++ app/front-end/src/types/enums/fileTypes.ts
@@ -1,0 +1,6 @@
+export enum FileTypes {
+  TXT = 'txt',
+  CSV = 'csv',
+  DOC = 'doc',
+  FOLDER = 'folder',
+}
