+++ app/front-end/src/types/enums/index.ts
@@ -1,0 +1,2 @@
+export { FileTypes } from './fileTypes';
+export { Colors } from './colors';
